code/3DI/PirateShooter/requirements.txt
autopep8
evdev
grpcio
grpcio-tools
protobuf